requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.0.0
numpy>=1.24.0

# MedImageInsight dependencies
mlflow==2.14.3
cffi==1.17.1
cloudpickle==3.0.0
colorama==0.4.6
einops==0.8.0
ftfy==6.2.3
fvcore==0.1.5.post20221221
mup==1.0.0
packaging>=24.1
pandas>=2.0.3
pyyaml>=6.0.2
requests>=2.32.3
sentencepiece==0.2.0
tenacity==9.0.0
timm==1.0.9
tornado==6.4.1

# For BLIP2
accelerate>=0.20.0

# Evaluation metrics
nltk>=3.8
evaluate>=0.4.0

# UI
gradio>=4.0.0

# Utilities
tqdm>=4.65.0
scikit-learn>=1.3.0
